+++ lib/Motor.h
@@ -12,7 +12,7 @@ void motor_init();
 void motor_foward(byte pwmR, byte pwmL);
 //void motor_foward_d(int distance); // distance(cm) メモリ食うので廃止
 //void motor_back(int pwm); //使わない
-void motor_rotate(int angle, float angleNow); //正面を0として±180度
+void motor_rotate(float angle, float angleNow); //正面を0として±180度
 void motor_stop();
 
 #endif
